// save-mate/src/app/core/enums/spending-categories-color.enum.ts
export enum SpendingCategoriesColor {
    HousingAndUtilities = "#B5C8EA",
    Transportation = "#F0A3B4",
    FoodAndHousehold = "#CC79A7",
    HealthAndPersonalExpenses = "#8F3985",
    ClothingAndOtherShopping = "#56B4E9",
    EntertainmentAndLeisure = "#0072B2",
    ChildrenPerFamily = "#F0E442",
    LoansAndDebts = "#E69F00",
    InsuranceAndFinancialProducts = "#D55E00",
    SavingAndInvesting = "#009E73",
    OtherPerVariableExpenses = "#7A9E7E"
}